Rename package to metricsreporterprometheus
Both metrics-reporter-jmx and metrics-reporter-prometheus plugins used
the same package name: com.googlesource.gerrit.plugins.metricsreporters
Adding a class with the same name in both plugins would be no problem
for the bazel build but it is a problem for the generated eclipse
project. Eclipse reports an error because of a type defined multiple
times.

Change-Id: I4c68e7c52df14ba9119c0bb82f2983487373dd1e