Add Metric containing Gerrit and Java version information
So far the Gerrit version could not be retrieved from Prometheus,
when using it to store metrics.

The version information can be send to Prometheus with a trick by
setting this information as label values in an otherwise static
metric. The labels can then be extracted and viewed in a table,
e.g. in Grafana. This allows to associate the Gerrit version with
the metrics measured at the same timeframe.

This change adds the gerrit_build_info metric with the labels:

- javaversion
- version (Gerrit version)
- revision (revision of Gerrit, if not using a release)

Change-Id: I4076c579fa4a137a5ce6c69777adb03eaa6de26f