Document how to specify multiple excludeMetrics
The documentation mentions multiple patterns can be given by repeating
the `excludeMetrics` setting but it was unclear to me how to proceed.

The setting is fetched using a getStringList() which comes from jgit and
indeed one just need to repeat it. An example makes it clearer.

The doc mentions the matching is done on the whole key, document that it
is internally using `regex.Matcher.matches()` (and not `lookingAt()`
which matches at the start of the subject).  The multiple values
example does use a lead carret '^' which indicates matching at the start
of metric name has to be set explicitly.

Change-Id: I89fd2d4bb6523446ba54f741767336f907c800ac