Add Metric containing current gerrit.config information
So far the gerrit.config could not be retrieved from Prometheus,
when using it to store metrics.

The gerrit.config information can be send to Prometheus with a trick by
setting this information as label values in an otherwise static
metric. The labels can then be extracted and viewed in a table,
e.g. in Grafana. This allows to associate the gerrit.config with
the metrics measured at the same timeframe.

This change adds the config_current metric with the label:
- gerrit_config

Only allow adminstrators to read the config metric since it may contain
sensible information.

Change-Id: Id21248d6d96ce23b69728d1ab90a789f53096746
Depends-On: I18ab5691c7b70e1413bfd6e589c3f0bd72720ed2